Refactor compile/run commands into two subcommands
