+++ bin/bondage
@@ -1,0 +1,1 @@
+../src/bin/bondage.js
